fix: Refactored build scripts.

+++ src/udt/listener_poller.cpp
@@ -39,7 +39,7 @@ int listener_poller<udt::epoll_poller>::poll (std::chrono::milliseconds millis, 
     if (n > 0) {
         for (UDTSOCKET u: _rep.readfds) {
             auto status = UDT::getsockstate(u);
-            LOGD(TAG, "UDT server socket state: {}", status);
+            LOGD(TAG, "UDT server socket state: {}", static_cast<int>(status));
 
             res++;
